promote svh calculation into its own directory
